+++ README.md
@@ -1,1 +1,1 @@
-# Калькулятор
+# Калькулятор 🔢
