+++ index.html
@@ -44,7 +44,7 @@
       </form>
     </fieldset>
   </main>
-  <script src="script.js"></script>
+  <script type="module" src="script.js"></script>
 </body>
 
 </html>
