+++ modules/Observable.mjs
@@ -1,0 +1,16 @@
+class Observable {
+  constructor() {
+    this.observers = [];
+  }
+  subscribe(f) {
+    this.observers.push(f);
+  }
+  unsubscribe(f) {
+    this.observers = this.observers.filter(subscriber => subscriber !== f);
+  }
+  notify(data) {
+    this.observers.forEach(observer => observer(data));
+  }
+}
+
+export {Observable};
