+++ style.css
@@ -1,7 +1,7 @@
 .body {
   height: 100vh;
   background-color: #424242;
-  background-image: url(./images/_.jpeg);
+  background-image: url(./images/fon.jpeg);
   background-repeat: no-repeat;
   background-position: center;
   background-size: cover;
